delete unused block executable type & move config functions to features
