+++ aptos-move/aptos-vm/src/move_vm_ext/vm.rs
@@ -21,12 +21,7 @@ use aptos_types::{
     on_chain_config::{FeatureFlag, Features, TimedFeatureFlag, TimedFeatures},
     transaction::user_transaction_context::UserTransactionContext,
 };
-use move_binary_format::{
-    deserializer::DeserializerConfig,
-    errors::VMResult,
-    file_format_common,
-    file_format_common::{IDENTIFIER_SIZE_MAX, LEGACY_IDENTIFIER_SIZE_MAX},
-};
+use move_binary_format::{deserializer::DeserializerConfig, errors::VMResult};
 use move_bytecode_verifier::VerifierConfig;
 use move_vm_runtime::{
     config::VMConfig, move_vm::MoveVM, native_extensions::NativeContextExtensions,
@@ -37,32 +32,6 @@ pub struct MoveVmExt {
     inner: MoveVM,
     chain_id: u8,
     features: Features,
-}
-
-pub fn get_max_binary_format_version(
-    features: &Features,
-    gas_feature_version_opt: Option<u64>,
-) -> u32 {
-    // For historical reasons, we support still < gas version 5, but if a new caller don't specify
-    // the gas version, we default to 5, which was introduced in late '22.
-    let gas_feature_version = gas_feature_version_opt.unwrap_or(5);
-    if gas_feature_version < 5 {
-        file_format_common::VERSION_5
-    } else if features.is_enabled(FeatureFlag::VM_BINARY_FORMAT_V7) {
-        file_format_common::VERSION_7
-    } else if features.is_enabled(FeatureFlag::VM_BINARY_FORMAT_V6) {
-        file_format_common::VERSION_6
-    } else {
-        file_format_common::VERSION_5
-    }
-}
-
-pub fn get_max_identifier_size(features: &Features) -> u64 {
-    if features.is_enabled(FeatureFlag::LIMIT_MAX_IDENTIFIER_LENGTH) {
-        IDENTIFIER_SIZE_MAX
-    } else {
-        LEGACY_IDENTIFIER_SIZE_MAX
-    }
 }
 
 impl MoveVmExt {
@@ -80,13 +49,9 @@ impl MoveVmExt {
     where
         F: Fn(DynamicExpression) + Send + Sync + 'static,
     {
-        // Note: binary format v6 adds a few new integer types and their corresponding instructions.
-        //       Therefore it depends on a new version of the gas schedule and cannot be allowed if
-        //       the gas schedule hasn't been updated yet.
         let max_binary_format_version =
-            get_max_binary_format_version(&features, Some(gas_feature_version));
-
-        let max_identifier_size = get_max_identifier_size(&features);
+            features.get_max_binary_format_version(Some(gas_feature_version));
+        let max_identifier_size = features.get_max_identifier_size();
 
         let enable_invariant_violation_check_in_swap_loc =
             !timed_features.is_enabled(TimedFeatureFlag::DisableInvariantViolationCheckInSwapLoc);
@@ -106,9 +71,9 @@ impl MoveVmExt {
 
         let mut builder = SafeNativeBuilder::new(
             gas_feature_version,
-            native_gas_params.clone(),
-            misc_gas_params.clone(),
-            timed_features.clone(),
+            native_gas_params,
+            misc_gas_params,
+            timed_features,
             features.clone(),
         );
